requirements.txt
flask
fastapi
uvicorn
requests
python-dotenv
supabase